Fix IDP options
