+++ app/subscriber/src/features/login/MobileLogin.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import { Button, Col, Row, Show, useKeycloakWrapper } from 'tno-core';
+import { Col, Row } from 'tno-core';
 
-import { Copyright } from './Copyright';
+import { IDPOptions } from './IDPOptions';
 import * as styled from './styled';
-import { SystemMessage } from './SystemMessage';
 
 export interface IMobileLoginProps {
   login: (hint?: string) => void;
@@ -15,59 +14,18 @@ export interface IMobileLoginProps {
  * @returns MobileLogin component.
  */
 export const MobileLogin: React.FC<IMobileLoginProps> = ({ login }) => {
-  const keycloak = useKeycloakWrapper();
-  const authority = keycloak.instance.authServerUrl?.replace(/\/$/, '') ?? window.location.href;
-  const isLocal =
-    new URL(authority).host.startsWith('localhost') ||
-    new URL(authority).host.startsWith('host.docker.internal');
-
   return (
     <styled.MobileLogin>
       <Col className="mobile-view">
         <Row className="mobile-title">
           <div className="app-logo"></div>
         </Row>
-        <Row className="containing-row">
-          <Col className="main-box">
-            <p className="top-bar-box">
-              Media Monitoring is a paid service offered through the BC Government that allows
-              subscribers to see British Columbia’s news at a glance.
-            </p>
-            <div className="containing-box centered-login-box">
-              <Col className="login-box">
-                <b>
-                  Login to your MMI account with one of the following accounts: IDIR, Microsoft
-                  Azure account, or custom email account:
-                </b>
-                <div>
-                  <div className="buttons">
-                    <Show visible={!isLocal}>
-                      <Button
-                        className="white idir-logo"
-                        onClick={() => login(isLocal ? 'gcpe-oidc' : 'idir')}
-                      >
-                        IDIR
-                      </Button>
-                    </Show>
-                    <Show visible={isLocal}>
-                      <Button className="white" onClick={() => login()}>
-                        Local
-                      </Button>
-                    </Show>
-                    <br />
-                    <p>
-                      <a href="mailto:Scott.Ryckman@gov.bc.ca">
-                        Learn more about obtaining a subscription
-                      </a>
-                    </p>
-                    <Copyright />
-                  </div>
-                </div>
-              </Col>
-            </div>
-            <SystemMessage />
-          </Col>
-        </Row>
+        <IDPOptions login={login}>
+          <br />
+          <p>
+            <a href="mailto:Scott.Ryckman@gov.bc.ca">Learn more about obtaining a subscription</a>
+          </p>
+        </IDPOptions>
       </Col>
     </styled.MobileLogin>
   );
